+++ src/types/types.ts
@@ -1369,6 +1369,16 @@ export interface EdgeCurrencyWallet {
   // Generic:
   readonly otherMethods: EdgeOtherMethods
 
+  readonly convertDenominatedToNative: (
+    denominatedAmount: string,
+    tokenId: EdgeTokenId
+  ) => Promise<string>
+
+  readonly convertNativeToDenominated: (
+    nativeAmount: string,
+    tokenId: EdgeTokenId
+  ) => Promise<string>
+
   /** @deprecated Use the information in EdgeCurrencyInfo / EdgeToken. */
   readonly denominationToNative: (
     denominatedAmount: string,