+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- added: Add tokenId-based denomination conversion utils to wallets
+- added: Add `payoutTokenId` to `EdgeTxSwap`
+
 ## 2.36.0 (2025-11-04)
 
 - added: Added `EdgeSubscribedAddress` type for `onSubscribeAddresses` and `startEngine`.