+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- added: Remove empty transaction metadata files when loading transaction data.
 - changed: Use syncRepo from upgraded edge-sync-client for syncing repo algorithm.
 
 ## 2.36.0 (2025-11-04)